implemented a condition to check the presence of hthe product before deleting it

+++ controller/deleteProduct.controller.js
@@ -3,13 +3,24 @@ const prisma = new PrismaClient()
 
 const deleteProduct = async(req, res) => {
     const productId = req.params.id
+
     try {
-        const deletingProduct =  await prisma.products.delete({
-            where:{
-                id : productId
+        const findTheProductToBeDeleted = await prisma.products.findUnique({
+            where: {
+                id: productId
             }
         })
-        res.status(200).json({ok:true, message: "The product was deleted successfully"})
+        if (!findTheProductToBeDeleted) {
+            res.status(200).json({ok:false, message: "The product to be deleted was not found"})
+        }else{
+            const deletingProduct =  await prisma.products.delete({
+                where:{
+                    id : productId
+                }
+            })
+            res.status(200).json({ok:true, message: "The product was deleted successfully"})
+        }
+        
     } catch (error) {
         res.status(500).json({ok:false, message:"Something is wrong"})
     }
